First pass at CPCP changes
This adds the foodcritc rules and some rubocop rules to help standardize
building this cookbook. These are only suggestions, and you are more
then welcome to edit how you think you should.

This is only the start of the changes that'll be made for CPCP, but it
is at least a good start.

Signed-off-by: JJ Asghar  <jj@chef.io>